# PERF: Use GEOSCoordSeq_copyFromBuffer for creating simple geometries