--- a/pygeos/_geometry.pyx
+++ b/pygeos/_geometry.pyx
@@ -12,6 +12,7 @@ import pygeos
 
 from pygeos._geos cimport (
     GEOSContextHandle_t,
+    GEOSCoordSeq_copyFromBuffer_r,
     GEOSCoordSeq_create_r,
     GEOSCoordSeq_destroy_r,
     GEOSCoordSeq_setX_r,
@@ -98,6 +99,7 @@ def simple_geometries_1d(object coordinates, object indices, int geometry_type, 
     cdef unsigned int dims = coordinates.shape[1]
     if dims not in {2, 3}:
         raise ValueError("coordinates should N by 2 or N by 3.")
+    cdef int has_z = int(dims == 3)
 
     if geometry_type not in {0, 1, 2}:
         raise ValueError(f"Invalid geometry_type: {geometry_type}.")
@@ -109,7 +111,7 @@ def simple_geometries_1d(object coordinates, object indices, int geometry_type, 
     if np.any(indices[1:] < indices[:indices.shape[0] - 1]):
         raise ValueError("The indices must be sorted.")  
 
-    cdef double[:, :] coord_view = coordinates
+    cdef double[:] coord_view = coordinates.ravel()
 
     # get the geometry count per collection (this raises on negative indices)
     cdef unsigned int[:] coord_counts = np.bincount(indices).astype(np.uint32)
@@ -134,37 +136,43 @@ def simple_geometries_1d(object coordinates, object indices, int geometry_type, 
                         f"Index {geom_idx} is missing from the input indices."
                     )
 
-            # check if we need to close a linearring
-            if geometry_type == 2:
-                ring_closure = 0
-                if geom_size == 3:
-                    ring_closure = 1
-                else:
-                    for coord_idx in range(dims):
-                        if coord_view[idx, coord_idx] != coord_view[idx + geom_size - 1, coord_idx]:
-                            ring_closure = 1
-                            break
-                # check the resulting size to prevent invalid rings
-                if geom_size + ring_closure < 4:
-                    # the error equals PGERR_LINEARRING_NCOORDS (in pygeos/src/geos.h)
-                    raise ValueError("A linearring requires at least 4 coordinates.")
+            # # check if we need to close a linearring
+            # if geometry_type == 2:
+            #     ring_closure = 0
+            #     if geom_size == 3:
+            #         ring_closure = 1
+            #     else:
+            #         for coord_idx in range(dims):
+            #             if coord_view[idx, coord_idx] != coord_view[idx + geom_size - 1, coord_idx]:
+            #                 ring_closure = 1
+            #                 break
+            #     # check the resulting size to prevent invalid rings
+            #     if geom_size + ring_closure < 4:
+            #         # the error equals PGERR_LINEARRING_NCOORDS (in pygeos/src/geos.h)
+            #         raise ValueError("A linearring requires at least 4 coordinates.")
 
-            seq = GEOSCoordSeq_create_r(geos_handle, geom_size + ring_closure, dims)
-            for coord_idx in range(geom_size):
-                if _set_xyz(geos_handle, seq, coord_idx, dims, coord_view, idx) == 0:
-                    GEOSCoordSeq_destroy_r(geos_handle, seq)
-                    return  # GEOSException is raised by get_geos_handle
-                idx += 1
+            # seq = GEOSCoordSeq_create_r(geos_handle, geom_size + ring_closure, dims)
+            # for coord_idx in range(geom_size):
+            #     if _set_xyz(geos_handle, seq, coord_idx, dims, coord_view, idx) == 0:
+            #         GEOSCoordSeq_destroy_r(geos_handle, seq)
+            #         return  # GEOSException is raised by get_geos_handle
+            #     idx += 1
+
+            seq = GEOSCoordSeq_copyFromBuffer_r(geos_handle, &coord_view[idx*dims], geom_size, has_z, 0)
+
+            if seq == NULL:
+                return  # GEOSException is raised by get_geos_handle
+            idx += geom_size
 
             if geometry_type == 0:
                 geom = GEOSGeom_createPoint_r(geos_handle, seq)
             elif geometry_type == 1:
                 geom = GEOSGeom_createLineString_r(geos_handle, seq)
             elif geometry_type == 2:
-                if ring_closure == 1:
-                    if _set_xyz(geos_handle, seq, geom_size, dims, coord_view, idx - geom_size) == 0:
-                        GEOSCoordSeq_destroy_r(geos_handle, seq)
-                        return  # GEOSException is raised by get_geos_handle
+                # if ring_closure == 1:
+                #     if _set_xyz(geos_handle, seq, geom_size, dims, coord_view, idx - geom_size) == 0:
+                #         GEOSCoordSeq_destroy_r(geos_handle, seq)
+                #         return  # GEOSException is raised by get_geos_handle
                 geom = GEOSGeom_createLinearRing_r(geos_handle, seq)
 
             if geom == NULL: