--- a/pygeos/_pygeos_api.pxd
+++ b/pygeos/_pygeos_api.pxd
@@ -30,6 +30,6 @@ cdef extern from "c_api.h":
     # they are declared that way in the header file).
     object PyGEOS_CreateGeometry(GEOSGeometry *ptr, GEOSContextHandle_t ctx)
     char PyGEOS_GetGEOSGeometry(PyObject *obj, GEOSGeometry **out) nogil
-    GEOSCoordSequence* PyGEOSCoordSeq_FromBuffer(GEOSContextHandle_t ctx, const double* buf,
+    GEOSCoordSequence* PyGEOS_CoordSeq_FromBuffer(GEOSContextHandle_t ctx, const double* buf,
                                                  unsigned int size, unsigned int dims,
                                                  char ring_closure) nogil