--- a/src/c_api.h
+++ b/src/c_api.h
@@ -32,8 +32,14 @@
 #define PyGEOS_GetGEOSGeometry_RETURN char
 #define PyGEOS_GetGEOSGeometry_PROTO (PyObject * obj, GEOSGeometry * *out)
 
+/* GEOSCoordSequence* PyGEOSCoordSeq_FromBuffer(GEOSContextHandle_t ctx, const double* buf,
+                                                unsigned int size, unsigned int dims, char ring_closure)*/
+#define PyGEOSCoordSeq_FromBuffer_NUM 2
+#define PyGEOSCoordSeq_FromBuffer_RETURN GEOSCoordSequence*
+#define PyGEOSCoordSeq_FromBuffer_PROTO (GEOSContextHandle_t ctx, const double* buf, unsigned int size, unsigned int dims, char ring_closure)
+
 /* Total number of C API pointers */
-#define PyGEOS_API_num_pointers 2
+#define PyGEOS_API_num_pointers 3
 
 #ifdef PyGEOS_API_Module
 /* This section is used when compiling pygeos.lib C extension.
@@ -42,6 +48,7 @@
 
 extern PyGEOS_CreateGeometry_RETURN PyGEOS_CreateGeometry PyGEOS_CreateGeometry_PROTO;
 extern PyGEOS_GetGEOSGeometry_RETURN PyGEOS_GetGEOSGeometry PyGEOS_GetGEOSGeometry_PROTO;
+extern PyGEOSCoordSeq_FromBuffer_RETURN PyGEOSCoordSeq_FromBuffer PyGEOSCoordSeq_FromBuffer_PROTO;
 
 #else
 /* This section is used in modules that use the PyGEOS C API
@@ -56,6 +63,9 @@ static void **PyGEOS_API;
 
 #define PyGEOS_GetGEOSGeometry \
     (*(PyGEOS_GetGEOSGeometry_RETURN(*) PyGEOS_GetGEOSGeometry_PROTO)PyGEOS_API[PyGEOS_GetGEOSGeometry_NUM])
+
+#define PyGEOSCoordSeq_FromBuffer \
+    (*(PyGEOSCoordSeq_FromBuffer_RETURN(*) PyGEOSCoordSeq_FromBuffer_PROTO)PyGEOS_API[PyGEOSCoordSeq_FromBuffer_NUM])
 
 /* Dynamically load C API from PyCapsule.
  * This MUST be called prior to using C API functions in other modules; otherwise