--- a/src/geos.c
+++ b/src/geos.c
@@ -4,8 +4,10 @@
 #include "geos.h"
 
 #include <Python.h>
+#include <numpy/ndarraytypes.h>
 #include <numpy/npy_math.h>
 #include <structmember.h>
+#include <stdio.h>
 
 /* This initializes a globally accessible GEOSException object */
 PyObject* geos_exception[1] = {NULL};
@@ -879,4 +881,56 @@ GEOSGeometry* PyGEOSForce2D(GEOSContextHandle_t ctx, GEOSGeometry* geom) {
 
 GEOSGeometry* PyGEOSForce3D(GEOSContextHandle_t ctx, GEOSGeometry* geom, double z) {
   return force_dims(ctx, geom, 3, z);
+}
+
+GEOSCoordSequence* PyGEOSCoordSeq_copyFromBuffer(GEOSContextHandle_t ctx, const double* buf,
+                                                 unsigned int size, unsigned int dims, char ring_closure,
+                                                 npy_intp cs1, npy_intp cs2) {
+  GEOSCoordSequence* coord_seq;
+  // unsigned int dims = 2 + hasZ;
+  int hasZ = dims == 3;
+  char* cp1;
+  char* cp2;
+  int i, j;
+  double first_coord, last_coord;
+  // printf("Shape: %d, %d \n", (int)size, (int)dims);
+  // printf("Steps: %d, %d \n", (int)cs1, (int)cs2);
+  // printf("Args: %d, %d\n", hasZ, ring_closure);
+  // printf("Args: %d, %d, %d\n", size, dims, ring_closure);
+
+#if GEOS_SINCE_3_10_0
+
+  if ((!ring_closure) && (cs1 == dims * 8) && (cs2 == 8)) {
+    printf("Going to call GEOSCoordSeq_copyFromBuffer_r\n");
+    coord_seq = GEOSCoordSeq_copyFromBuffer_r(ctx, buf, size, hasZ, 0);
+    return coord_seq;
+  }
+
+#endif
+
+  coord_seq = GEOSCoordSeq_create_r(ctx, size + ring_closure, dims);
+  if (coord_seq == NULL) {
+    return NULL;
+  }
+  cp1 = (char*)buf;
+  for (i = 0; i < size; i++, cp1 += cs1) {
+    cp2 = cp1;
+    for (j = 0; j < dims; j++, cp2 += cs2) {
+      if (!GEOSCoordSeq_setOrdinate_r(ctx, coord_seq, i, j, *(double*)cp2)) {
+        GEOSCoordSeq_destroy_r(ctx, coord_seq);
+        return NULL;
+      }
+    }
+  }
+  /* add the closing coordinate if necessary */
+  if (ring_closure) {
+    for (j = 0; j < dims; j++){
+      first_coord = *(double*)((char*)buf + j * cs2);
+      if (!GEOSCoordSeq_setOrdinate_r(ctx, coord_seq, size, j, first_coord)) {
+        GEOSCoordSeq_destroy_r(ctx, coord_seq);
+        return NULL;
+      }
+    }
+  }
+  return coord_seq;
 }