… CHANGELOG.rst
… pygeos/tests/test_creation.py
… pygeos/tests/test_creation_indices.py
… src/ufuncs.c