Experiment in using the new `GEOSCoordSeq_copyFromBuffer` for creating CoordinateSequences instead of setting each ordinate individually.

This requires GEOS 3.10, so will require https://github.com/pygeos/pygeos/issues/255 to use this in cython.

I was benchmarking it with the US census polygons ([tl_2019_us_zcta510.zip](https://catalog.data.gov/dataset/tiger-line-shapefile-2019-2010-nation-u-s-2010-census-5-digit-zip-code-tabulation-area-zcta5-na), 33144 (multi)polygons, with on average ~1200 coordinates per ring).

Focusing on LinearRing creation (the actual creation of (Multi)Polygons from those rings will of course see less of a difference):

I saved the coordinates and offsets in a numpy file, to load them again to benchmark (see https://nbviewer.org/gist/jorisvandenbossche/4fab9c94525dcb5ddda8a89b4a5ad9f3 for the full notebook):

```python
import numpy as np
import pygeos

with np.load("pygeos-benchmark-linearrings.npz") as data:
    offsets = data["offsets"]
    coords = data["coords"]

ring_lengths = np.diff((offsets / 2).astype(int))
ring_indices = np.repeat(np.arange(len(ring_lengths)), ring_lengths)
rings = pygeos.linearrings(coords, indices=ring_indices)
```

and this gives:

```
In [2]: %timeit pygeos.linearrings(coords, indices=ring_indices)
1.13 s ± 6.86 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)   # <-- master
558 ms ± 2.35 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <-- PR
```

I had maybe hoped on a bit more :), but it's certainly a nice speed-up.