@jorisvandenbossche 2021-11-18
Doing a profile of the above benchmark with py-spy gives the following:

![pygeos-linearrings](https://user-images.githubusercontent.com/1020496/142408961-75425d22-0189-4930-acd2-196916ecb712.png)

So this confirms that inside `pygeos.linearrings` now most time is spent in `GEOSCoordSeq_copyFromBuffer` (66%). Another 25% is from `np.bincount`, and another 8% is from the comparison `indices[1:] < indices[:indices.shape[0] - 1]`(validating the indices are sorted)