@brendan-ward 2021-11-18
Thanks for working on this!  The results are promising (though a bummer that we spend so much time just validating inputs).

I wonder if it would be easier to introduce this first in the ufuncs (e.g., `lib.linestrings`) where we can use the C macros to guard the GEOS version?  That would unblock the dependency on adding GEOS version control mechanisms to Cython in order to first start using `GEOSCoordSeq_copyFromBuffer` in pygeos?  That said, we'll certainly want this in Cython too, just was trying to think through smaller steps to bringing this in.