@jorisvandenbossche 2021-11-18
> I wonder if it would be easier to introduce this first in the ufuncs (e.g., `lib.linestrings`) where we can use the C macros to guard the GEOS version?

That would indeed certainly be _easier_, but in practice those ufuncs are not super useful for general IO (except for points), since you almost never have all polygons of the same size (eg I couldn't have done my benchmark with real world data as I did above). So that's the reason I wanted to start looking at the non-ufunc versions. 
(but we certainly want to add it to the ufuncs as well)