@jorisvandenbossche 2021-11-18
> though a bummer that we spend so much time just validating inputs

It's not actually only validation, but also the calculation of the size of each CoordSeq that takes quite a bit of time. That's a consequence of how we expect the indices to be laid out (as `[0, 0, 0, 0, 1, 1, 2, 2, 2]` instead of `[0, 4, 6, 9]`). 
Long term, I think it would be nice to also have a routine that takes that kind of offsets (since this is what Arrow will natively store). 

The validation indeed also takes some time. We could maybe add a keyword to disable this, if you already know for sure that this is the case (but with a default to always validate).