@brendan-ward 2021-11-18
Likely crazy idea, not yet thought through: what if we exposed a function from pygeos C API that we consume in Cython so that it takes the same inputs as are now being passed into `GEOSCoordSeq_copyFromBuffer`.  Then, instead of using this function direction from GEOS, we call back into pygeos C API, which then uses C Macros to version-guard the underlying GEOS functions used.  I'm thinking that what we add would be lightweight wrappers around some of these important low-level GEOS functions where they are only available in certain recent versions of GEOS.

I.e., instead of adding shims to Cython build system to conditionally make certain GEOS functions available to Cython, we provide Cython C function shims that are implemented back in C on our end.

I don't know yet how the complexity of that would differ vs just making the Cython build system GEOS version aware.