@jorisvandenbossche 2021-11-18
That's not so crazy I think :)

We could indeed have a C function that either is a simple call of `GEOSCoordSeq_copyFromBuffer` or either creates a CoordSeq manually and fills this in iteratively, depending on the GEOS version (which is easy to do in C). And then we can call that in the cython code. This is basically the `_set_xyz` + the loop in which it is called that we now do custom in cython. 

That might indeed end up simpler.