@caspervdw 2021-11-19
Thanks for investigating this and doing the benchmarks! Interesting results. For sure the new GEOS function gives significant improvements. And we have some handles to optimize even further.

We could indeed make an optional “offsets” keyword argument that is preferred above the “indices”. It depends on your use case which representation is fastest, apparently here “offsets” is a better match. This is material for another issue I guess.

For the coordinates -> coordseq conversion I think the proposed solution of @brendan-ward is not crazy at all. I think that is the way to go.