@jorisvandenbossche 2021-11-21
I did a first attempt to write a generic "create and fill in coord seq" function that can be used in in the ufuncs (and then later in the cython code as well). Some complications: the ufuncs can also handle differently strided arrays (so I have to check if the steps are correct to use the copyFromBuffer), and needed some casts back and forth between `*double` and `*char` (the steps are expressed in bytes, not for `* double`). This can probably be cleaned up a bit more.