@jorisvandenbossche 2021-11-21
Quick performance check for the ufunc as well (not compared with master, but comparing with itself but triggering the copyFromBuffer path by providing unclosed vs closed ring coordinates):

```python
In [15]: coords1 = np.random.randn(10000, 100, 2)

In [16]: %timeit pygeos.linearrings(coords1)
17.2 ms ± 603 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [17]: coords1[:, [99], :] = coords1[: , [0], :]

In [18]: %timeit pygeos.linearrings(coords1)   # <-- closed rings, uses the copyFromBuffer fast path
8.93 ms ± 598 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

and with fewer but larger rings:

```python
In [19]: coords1 = np.random.randn(100, 10000, 2)

In [20]: %timeit pygeos.linearrings(coords1)
10.8 ms ± 111 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [21]: coords1[:, [9999], :] = coords1[: , [0], :]

In [22]: %timeit pygeos.linearrings(coords1)
3.54 ms ± 44.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```