@caspervdw 2021-11-29
Looks good to merge! Can you add a line to the changelog?